refactor wait into Waiter interface
